feat: add distance, centroids

+++ K-MEANS/src/kmeans/KMeans.java
@@ -93,7 +93,7 @@ public class KMeans {
         atractores.add(new Coordenadas(x, y, c, nC));
     }
     
-    public float euclidiana(int x1, int y1, int x2, int y2){
+    public float euclidiana(float x1, float y1, float x2, float y2){
         float d;
         return d =(float) Math.sqrt( Math.pow(x2-x1,2)+ Math.pow(y2-y1,2));
     }
